# Avoid method invalidations when CxxWrap is loaded