In the pretty printing code, there are many invocations of ==(::Any,::Symbol)
due to the left side not being inferable (be it due to `@nospecialize`, or
because `obj.args[1]` can be anything, or other reasons).

To give an idea of the effect, consider the following timings, made with the
latest Julia master branch, which now contains code for caching compiled code.
I've benchmarked a simple Hecke computation, with and without loading CxxWrap:

Without CxxWrap:

    julia> using Hecke ; @time @eval maximal_order(quadratic_field(-3)[1]);
      0.029069 seconds (74.71 k allocations: 4.971 MiB, 121.79% compilation time)

With CxxWrap but without this patch:

    julia> using Hecke ; using CxxWrap ; @time @eval maximal_order(quadratic_field(-3)[1]);
      3.239503 seconds (3.83 M allocations: 255.878 MiB, 1.47% gc time, 114.90% compilation time: 86% of which was recompilation)

    julia> using CxxWrap; using Hecke ; @time @eval maximal_order(quadratic_field(-3)[1]);
      2.834618 seconds (3.52 M allocations: 236.457 MiB, 2.08% gc time, 114.91% compilation time: 86% of which was recompilation)

With CxxWrap and with this patch:

    julia> using Hecke ; using CxxWrap ; @time @eval maximal_order(quadratic_field(-3)[1]);
      2.002340 seconds (2.61 M allocations: 175.162 MiB, 1.93% gc time, 117.60% compilation time: 84% of which was recompilation)

    julia> using CxxWrap; using Hecke ; @time @eval maximal_order(quadratic_field(-3)[1]);
      1.756986 seconds (2.33 M allocations: 157.248 MiB, 2.46% gc time, 116.21% compilation time: 85% of which was recompilation)